Create NumberOfFlipsToMakeAORBEqualToC.java
We have three positive numbers, a, b, and c. We have to find the minimum flips required in some bits of a, b to make (a OR (b == c)). Here we are considering Bitwise OR operation. The flip operation consists of changing any single bit 1 to 0 or changing the bit 0 to 1 in their binary representation.